Change handle type back to int

+++ GOESP/GameData.h
@@ -85,7 +85,7 @@ struct ProjectileData : BaseData {
     ProjectileData(Entity* projectile) noexcept;
     void update(Entity* projectile) noexcept;
 
-    constexpr auto operator==(uintptr_t otherHandle) const noexcept
+    constexpr auto operator==(int otherHandle) const noexcept
     {
         return handle == otherHandle;
     }
@@ -93,7 +93,7 @@ struct ProjectileData : BaseData {
     bool exploded = false;
     bool thrownByLocalPlayer = false;
     bool thrownByEnemy = false;
-    uintptr_t handle;
+    int handle;
     const char* name = nullptr;
     std::vector<std::pair<float, Vector>> trajectory;
 };
